+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - add option to set psql output format to HTML
 - allow package to be called programmatically
+- add custom flag parser for SQL flags
 
 ## v0.1.0 (2019-06-28)